agregue plane.h y transform.h pero no estan implementados solo para que no me tire error el scene_parser.cpp

+++ src/main.cpp
@@ -55,10 +55,10 @@ void parseArguments(int argc, char* argv[]) {
         } else if (!strcmp(argv[i],"-depth")) {
             i++;
             assert (i < argc);
-            depth_min = atof(argv[i]);
+            depth_min = (float)atof(argv[i]);
             i++;
             assert (i < argc);
-            depth_max = atof(argv[i]);
+            depth_max = (float)atof(argv[i]);
             i++;
             assert (i <= argc);
             depth_file = argv[i];
